Add MOSFET level shifter and split v1.8 rail
The MOSFET provides level shifting for the RTS line. The RTS line
provides the reset signal for the Atmel processor.

The 1.8V rail needed to be split so that the FT230X chip works correctly
regardless of the order the board is powered up. Without it the board
would need to be powered up before the USB cable is plugged in.

Signed-off-by: Grant Likely <grant.likely@linaro.org>