Switch transistor schematic symbol to BC817
The part is still an S8050, but we want to use the SOT-23 footprint
instead of TO-92. The default Kicad libraries don't have a SOT-23
version of the S8050, but it does have one for the BC817 which uses the
same pinout. Switch the schematic symbol so that KiCad stops complaining
about the footprint being different and needing to be "rescued"

Signed-off-by: Grant Likely <grant.likely@linaro.org>