Rewire PB5 LED to drive high to light and use a 1K resistor
The first rev had the LED tied to +5V, and PB5 needed to pull it low to
light. This doesn't play well with USBTinyISP because it puts too much
load on the line for the programmer to actually program the Arduino, and
it is opposite to the way the led is wired on every other Arduino. Fix
it to be consistent with the rest of the Arduino world.

Signed-off-by: Grant Likely <grant.likely@linaro.org>