Remove i2c pullups and add 22 ohm series resistors to reduce noise
There have been all kinds of problems with noise on the i2c bus, but the
pull up strength seems to be just fine for the bus speed that we're
using. Remove the pull up resistors, but put in series resistors to
reduce the effects of crosstalk on the i2c lines.

Also, split the data and clock signals so they don't run parallel across
the whole board. That should also reduce noise.

Signed-off-by: Grant Likely <grant.likely@linaro.org>