Replace TXS0108 level shifters for i2c with PCA9306
The TXS0108 doesn't work well with long trace lengths on i2c. When there
is too much capacitive load on the line the integrated one-shot drivers
go into oscillation and it locks up the bus. Since we cannot make the
TXS0108 work reliably, use a different part on the i2c lines. The
PCA9306 is a simpler circuit without the one-shot driver and it works
reliably on a prototyped bread-board circuit.

Sadly, the PCA9306 takes up a lot more room because it doesn't have
integrated pull-up resistors, and there wasn't any more room on the
board to place the new parts. Drop the switchable voltage levels to get
back some space at the top of the board where the shifters need to be.
Instead, hardwire each I2C bus at a different voltage level. I2C0 is
shifted to 5V, and I2C1 to 3.3V. Fix the ATMEGA to 5V only. GPIO-A and
GPIO-B are shifted to 5V, and GPIOs E-L are shifted to 3.3V (5V
tolerant). We don't use GPIOs C & D.

Finally, the 96B GPIOs and I2C grove connectors have swapped sides to
make routing easier.

Signed-off-by: Grant Likely <grant.likely@linaro.org>