Protect baseboard if Sensors board is installed incorrectly
Disconnect pins 36 (PWR_IN) and 40 (GND) to protect the baseboard if P1
gets installed off-by-one. Keeping PWR_IN pins 36 and 38 tied together
will short +12V directly to GPIO_L (34) or GND (40) if the connector was
off by one. Having pin 40 attached to GND would cause +12V to be shorted
directly to the +5V rail. To protect the baseboard, remove these direct
connections. Using only one PWR_IN pin is more than sufficient for this
board because it isn't used for anything. There are also three other GND
pins on the connector rated at 2A each. Dropping one should not be an
issue.

These changes don't guarantee to protect the baseboard. 5V would still
get shunted to the 1.8V rail which may still destroy the board, but
there is at least a fighting chance of not frying something.

Things would be so much better if we had a keyed connector

Signed-off-by: Grant Likely <grant.likely@linaro.org>